Moved Listener to its own place. Created res folder to store all this constants aswell exceptions and config values. Included listeners in Flask Constructor
