+++ src/index.js
@@ -1,14 +1,9 @@
 import Flask from './Flask'
-import {
-  GLOBAL_LISTENER_NAME,
-  RESOLVED_LISTENER_NAME
-} from './util/constants'
+import * as listeners from './res/listeners'
 
 Flask._version = '1.0.3'
 Flask._env = 'development'
 
+Flask.listeners = listeners
+
 export default Flask
-export {
-  GLOBAL_LISTENER_NAME,
-  RESOLVED_LISTENER_NAME
-}
