+++ src/res/config.js
@@ -1,0 +1,2 @@
+export const PARAMETER_DELIMITER_CHAR = '%'
+export const SERVICE_DELIMITER_CHAR = '@'
