+++ src/util/config.js
@@ -1,8 +1,5 @@
 import Reflect from 'harmony-reflect'
-import {
-  PARAMETER_DELIMITER_CHAR,
-  SERVICE_DELIMITER_CHAR
-} from './constants'
+import { PARAMETER_DELIMITER_CHAR, SERVICE_DELIMITER_CHAR } from '../res/config'
 
 export const configureFlask = (configObject, flask) => {
   registerConfigValues(Reflect.get(configObject, 'config') || {}, flask)
