--- src/util/constants.js
@@ -1,4 +1,0 @@
-export const PARAMETER_DELIMITER_CHAR = '%'
-export const SERVICE_DELIMITER_CHAR = '@'
-export const GLOBAL_LISTENER_NAME = '__global__'
-export const RESOLVED_LISTENER_NAME = 'resolved'
