+++ test/InitializationSpec.js
@@ -1,17 +1,15 @@
 import Flask from '../src/Flask'
 import { Service, Singleton, Parameter } from '../src/services'
-import {
-  GLOBAL_LISTENER_NAME,
-  RESOLVED_LISTENER_NAME
-} from '../src/util/constants'
+import { GLOBAL_LISTENER_NAME, ON_RESOLVED } from '../src/res/listeners'
+import { PARAMETER_DELIMITER_CHAR, SERVICE_DELIMITER_CHAR } from '../src/res/config'
 
 describe('Initial Configuration', () => {
   it('Load default config values if config values not present', () => {
     const flask = new Flask()
     assert.property(flask.config, 'serviceDelimiter')
     assert.property(flask.config, 'paramDelimiter')
-    assert.equal(flask.config.serviceDelimiter, '@')
-    assert.equal(flask.config.paramDelimiter, '%')
+    assert.equal(flask.config.serviceDelimiter, SERVICE_DELIMITER_CHAR)
+    assert.equal(flask.config.paramDelimiter, PARAMETER_DELIMITER_CHAR)
   })
 
   it('Override config values from initial config', () => {
@@ -155,17 +153,17 @@ describe('Initial Configuration', () => {
     const func3 = () => {}
     const config = {
       listeners: {
-        [RESOLVED_LISTENER_NAME]: [func1],
+        [ON_RESOLVED]: [func1],
         custom: [func2, func3]
       }
     }
 
     const flask = new Flask(config)
-    assert.property(flask.listeners, RESOLVED_LISTENER_NAME)
+    assert.property(flask.listeners, ON_RESOLVED)
     assert.property(flask.listeners, 'custom')
-    assert.property(flask.listeners[RESOLVED_LISTENER_NAME], GLOBAL_LISTENER_NAME)
+    assert.property(flask.listeners[ON_RESOLVED], GLOBAL_LISTENER_NAME)
     assert.property(flask.listeners['custom'], GLOBAL_LISTENER_NAME)
-    assert.deepEqual(flask.listeners[RESOLVED_LISTENER_NAME][GLOBAL_LISTENER_NAME], [func1])
+    assert.deepEqual(flask.listeners[ON_RESOLVED][GLOBAL_LISTENER_NAME], [func1])
     assert.deepEqual(flask.listeners['custom'][GLOBAL_LISTENER_NAME], [func2, func3])
   })
 
@@ -179,7 +177,7 @@ describe('Initial Configuration', () => {
         aliasA: {
           service: serviceA,
           listeners: {
-            [RESOLVED_LISTENER_NAME]: [func1],
+            [ON_RESOLVED]: [func1],
             custom: [func2, func3]
           }
         }
@@ -188,11 +186,11 @@ describe('Initial Configuration', () => {
 
     const flask = new Flask(config)
     const service = flask.services[0]
-    assert.property(flask.listeners, RESOLVED_LISTENER_NAME)
+    assert.property(flask.listeners, ON_RESOLVED)
     assert.property(flask.listeners, 'custom')
-    assert.property(flask.listeners[RESOLVED_LISTENER_NAME], service.alias)
+    assert.property(flask.listeners[ON_RESOLVED], service.alias)
     assert.property(flask.listeners['custom'], service.alias)
-    assert.deepEqual(flask.listeners[RESOLVED_LISTENER_NAME][service.alias], [func1])
+    assert.deepEqual(flask.listeners[ON_RESOLVED][service.alias], [func1])
     assert.deepEqual(flask.listeners['custom'][service.alias], [func2, func3])
   })
 })
